add rules store data barang masuk keluar

+++ routes/web.php
@@ -32,11 +32,11 @@ Route::get('/', function () {
     if (Auth::check()) {
         // return Response::json(Auth::user()->role);
         if (Auth::user()->role == 'admin') {
-            return redirect('/listbarang');
+            return redirect('/admin/listbarang');
         }elseif (Auth::user()->role == 'logistic') {
             return redirect('/logistik/barang');
         }elseif (Auth::user()->role == 'hr') {
-            return redirect('/proyek');
+            return redirect('/hr/proyek');
         }elseif (Auth::user()->role == 'mandor') {
             return redirect('/request');
         }
@@ -61,22 +61,27 @@ Route::get('/', function () {
 
 Route::group(['middleware' => ['auth:sanctum', 'checkRole:admin']], function()
 {
-    Route::get('/listbarang', [BarangController::Class, 'index']);
-    Route::post('/listbarang', [BarangController::Class, 'store']);
-    Route::put('/listbarang/{id}', [BarangController::Class, 'update']);
-    Route::delete('/listbarang/{id}', [BarangController::Class, 'destroy']);
-    Route::get('/listtenagakerja', [TenagaKerjaController::Class, 'index']);
-    Route::post('/listtenagakerja', [TenagaKerjaController::Class, 'store']);
-    Route::put('/listtenagakerja/{id}', [TenagaKerjaController::Class, 'update']);
-    Route::delete('/listtenagakerja/{id}', [TenagaKerjaController::Class, 'destroy']);
-    Route::get('/listtenagakerja/show/{id}', [TenagaKerjaController::Class, 'show']);
-    Route::get('/showdetailbarang/{id}', [BarangController::Class, 'showDetail']);
-    Route::put('/detail/{id}', [BarangController::Class, 'detailBarang']);
-    Route::get('/listproyek', [ProjectController::Class, 'index']);
-    Route::get('/listproyek/create', [ProjectController::Class, 'create']);
-    Route::post('/listproyek', [ProjectController::Class, 'store']);
-    Route::put('/listproyek/{id}', [ProjectController::Class, 'update']);
-    Route::delete('/listproyek/{id}', [ProjectController::Class, 'destroy']);
+    Route::get('/admin/listbarang', [BarangController::Class, 'index']);
+    Route::post('/admin/listbarang', [BarangController::Class, 'store']);
+    Route::put('/admin/listbarang/{id}', [BarangController::Class, 'update']);
+    Route::delete('/admin/listbarang/{id}', [BarangController::Class, 'destroy']);
+    Route::get('/admin/showdetailbarang/{id}', [BarangController::Class, 'showDetail']);
+    Route::get('/admin/detailbarang/{id}/{start}/{end}/{tp}', [BarangController::Class, 'detailBarang']);
+
+    Route::get('/admin/listtenagakerja', [TenagaKerjaController::Class, 'index']);
+    Route::post('/admin/listtenagakerja', [TenagaKerjaController::Class, 'store']);
+    Route::put('/admin/listtenagakerja/{id}', [TenagaKerjaController::Class, 'update']);
+    Route::delete('/admin/listtenagakerja/{id}', [TenagaKerjaController::Class, 'destroy']);
+    Route::get('/admin/listtenagakerja/show/{id}', [TenagaKerjaController::Class, 'show']);
+
+    Route::get('/admin/listproyek', [ProjectController::Class, 'index']);
+    Route::get('/admin/listproyek/create', [ProjectController::Class, 'create']);
+    Route::post('/admin/listproyek', [ProjectController::Class, 'store']);
+    Route::put('/admin/listproyek/{id}', [ProjectController::Class, 'update']);
+    Route::delete('/admin/listproyek/{id}', [ProjectController::Class, 'destroy']);
+    Route::get('/admin/proyek/show/{id}', [ProjectController::Class, 'show']);
+    Route::put('/admin/proyek/detail/{id}/{start}/{end}', [ProjectController::Class, 'detail']);
+    Route::put('/admin/proyek/alocated/{id}/{start}/{end}', [ProjectController::Class, 'sumBarang']);
     // Route::resource('tenagakerja', TenagaKerjaController::Class);
     // Route::resource('baranginout', ToolsInOutController::Class);
     Route::get('/listbaranginout', [BarangInOutController::Class, 'index']);
@@ -117,20 +122,20 @@ Route::group(['middleware' => ['auth:sanctum', 'checkRole:logistic']], function(
 
 Route::group(['middleware' => ['auth:sanctum', 'checkRole:hr']], function()
 {
-    Route::get('/proyek', [ProjectController::Class, 'index']);
-    Route::get('/proyek/create', [ProjectController::Class, 'create']);
-    Route::post('/proyek', [ProjectController::Class, 'store']);
-    Route::put('/proyek/{id}', [ProjectController::Class, 'update']);
-    Route::get('/proyek/show/{id}', [ProjectController::Class, 'show']);
-    Route::put('/proyek/detail/{id}/{start}/{end}', [ProjectController::Class, 'detail']);
-    Route::put('/proyek/alocated/{id}/{start}/{end}', [ProjectController::Class, 'sumBarang']);
-    Route::delete('/proyek/{id}', [ProjectController::Class, 'destroy']);
+    Route::get('/hr/proyek', [ProjectController::Class, 'index']);
+    Route::get('/hr/proyek/create', [ProjectController::Class, 'create']);
+    Route::post('/hr/proyek', [ProjectController::Class, 'store']);
+    Route::put('/hr/proyek/{id}', [ProjectController::Class, 'update']);
+    Route::get('/hr/proyek/show/{id}', [ProjectController::Class, 'show']);
+    Route::put('/hr/proyek/detail/{id}/{start}/{end}', [ProjectController::Class, 'detail']);
+    Route::put('/hr/proyek/alocated/{id}/{start}/{end}', [ProjectController::Class, 'sumBarang']);
+    Route::delete('/hr/proyek/{id}', [ProjectController::Class, 'destroy']);
     // Route::resource('proyek', ProjectController::Class);
-    Route::get('/tenagakerja', [TenagaKerjaController::Class, 'index']);
-    Route::post('/tenagakerja', [TenagaKerjaController::Class, 'store']);
-    Route::get('/tenagakerja/show/{id}', [TenagaKerjaController::Class, 'show']);
-    Route::put('/tenagakerja/{id}', [TenagaKerjaController::Class, 'update']);
-    Route::delete('/tenagakerja/{id}', [TenagaKerjaController::Class, 'destroy']);
+    Route::get('/hr/tenagakerja', [TenagaKerjaController::Class, 'index']);
+    Route::post('/hr/tenagakerja', [TenagaKerjaController::Class, 'store']);
+    Route::get('/hr/tenagakerja/show/{id}', [TenagaKerjaController::Class, 'show']);
+    Route::put('/hr/tenagakerja/{id}', [TenagaKerjaController::Class, 'update']);
+    Route::delete('/hr/tenagakerja/{id}', [TenagaKerjaController::Class, 'destroy']);
 });
 
 Route::group(['middleware' => ['auth:sanctum', 'checkRole:mandor']], function()
